refactor: album - reader, writer
