+++ src/module/album/album.module.ts
@@ -1,13 +1,14 @@
 import { Module } from '@nestjs/common';
 import { AlbumController } from './album.controller';
 import { AlbumService } from './album.service';
-import { AlbumRepository } from './repository/album.repository';
+import { AlbumWriter } from './repository/album.writer';
+import { AlbumReader } from './repository/album.reader';
 import { FeedModule } from '../feed/feed.module';
 
 @Module({
   imports: [FeedModule],
   controllers: [AlbumController],
-  providers: [AlbumService, AlbumRepository],
-  exports: [AlbumRepository],
+  providers: [AlbumService, AlbumWriter, AlbumReader],
+  exports: [AlbumWriter, AlbumReader],
 })
 export class AlbumModule {}
