+++ src/module/album/repository/album.writer.ts
@@ -1,25 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { Prisma } from '@prisma/client';
-import { kyselyUuid, prismaUuid } from 'src/shared/util/convert-uuid';
+import { prismaUuid } from 'src/shared/util/convert-uuid';
 import { TransactionHost } from '@nestjs-cls/transactional';
 import { TransactionalAdapterPrisma } from '@nestjs-cls/transactional-adapter-prisma';
 
 @Injectable()
-export class AlbumRepository {
+export class AlbumWriter {
   constructor(
     private readonly txHost: TransactionHost<TransactionalAdapterPrisma>,
   ) {}
-
-  async findManyByUserId(userId: string) {
-    return await this.txHost.tx.album.findMany({
-      where: {
-        userId,
-      },
-      orderBy: {
-        order: 'asc',
-      },
-    });
-  }
 
   async create(
     userId: string,
@@ -73,33 +62,5 @@ export class AlbumRepository {
       AND "userId" = ${prismaUuid(userId)}
     `;
     return;
-  }
-
-  async findOneById(id: string) {
-    return await this.txHost.tx.album.findUnique({
-      where: { id },
-    });
-  }
-
-  async findManyWithCountByUserId(userId: string) {
-    const albums = await this.txHost.tx.$kysely
-      .selectFrom('Album')
-      .where('Album.userId', '=', kyselyUuid(userId))
-      .select((eb) =>
-        eb
-          .selectFrom('Feed')
-          .whereRef('Feed.albumId', '=', 'Album.id')
-          .select((eb) => eb.fn.count<bigint>('Feed.id').as('feedCount'))
-          .as('feedCount'),
-      )
-      .select(['Album.id', 'Album.name'])
-      .orderBy('Album.order', 'asc')
-      .execute();
-
-    return albums.map((album) => ({
-      id: album.id,
-      name: album.name,
-      feedCount: album.feedCount ? Number(album.feedCount) : 0,
-    }));
   }
 }
